+++ client/src/components/Track.js
@@ -3,20 +3,21 @@ import React from "react"
 
 export default function Track({title, artist_names, bpm, audio, onDeleteTrack, id}) {
     const handleDeleteTrack = (e) => {
+        if(window.confirm("Are you sure you want to delete this track?")){ 
         fetch(`/tracks/${id}`,{
           method:"DELETE"
         })
         .then(() => {
           onDeleteTrack(id)
         })
-    }
-
-
+    }}
 
     return (
         <div className="item" style={{padding: "10px"}}>
             <div className="content">
-                <div><h5>{title}</h5></div>
+                <div>
+                    <h5>{title}</h5>
+                </div>
                 <div className="meta">
                     <span>{artist_names}</span>
                     <span>{bpm} bpm</span>
@@ -25,11 +26,10 @@ export default function Track({title, artist_names, bpm, audio, onDeleteTrack, i
                     <div>AUDIO FILE GOES HERE{audio}</div>
                 </div>
                 <div className="actions tiny">
-                <a onClick={handleDeleteTrack} className="delete">Delete Track</a>
+                    <a className="delete">Edit  </a>
+                    <a onClick={handleDeleteTrack} className="delete">  Delete</a>
+                </div>
             </div>
-            </div>
-
-
         </div>
     )
